// DorotOlam/xbid_2/src/app/models/profilepage.model.ts
//export class profileDetails{
//    birthDate: Date;
//    birthplace: string="";
//    deathdate: Date;
//    deathage: number;
//    deathplace: string="";
//    burialplace: string="";
//    lifehistory: string="";
//    residence: string="";
//    tags:any="";
//    moreAddress:MoreAddress[]=[];
//}
//export class MoreAddress{
//name:string;
//}

export class profileDetails{
    birthDate: Date;
    birthplace: string="";
    deathdate: Date;
    deathage: number;
    deathplace: string="";
    burialplace: string="";
    lifehistory: string="";
    residence: string="";
    tags:any="";
    moreAddress:MoreAddress[]=[];
}
export class MoreAddress{
name:string;
}